Revert "[WFCORE-860] : Use notifications to help the scanner get a proper status of all deployments between scans"
