+++ deployment-scanner/src/main/java/org/jboss/as/server/deployment/scanner/FileSystemDeploymentService.java
@@ -27,8 +27,6 @@ import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.CAN
 import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.COMPOSITE;
 import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.CONTENT;
 import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.DEPLOYMENT;
-import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.DEPLOYMENT_DEPLOYED_NOTIFICATION;
-import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.DEPLOYMENT_UNDEPLOYED_NOTIFICATION;
 import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.ENABLED;
 import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.FAILED;
 import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.FAILURE_DESCRIPTION;
@@ -54,8 +52,6 @@ import java.io.FileFilter;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.net.MalformedURLException;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -76,8 +72,6 @@ import org.jboss.as.controller.ControlledProcessState;
 import org.jboss.as.controller.ControlledProcessStateService;
 import org.jboss.as.controller.OperationFailedException;
 import org.jboss.as.controller.PathAddress;
-import org.jboss.as.controller.notification.Notification;
-import org.jboss.as.controller.notification.NotificationHandler;
 import org.jboss.as.controller.operations.common.Util;
 import org.jboss.as.server.deployment.DeploymentAddHandler;
 import org.jboss.as.server.deployment.DeploymentDeployHandler;
@@ -97,7 +91,7 @@ import org.jboss.dmr.Property;
  *
  * @author Brian Stansberry
  */
-class FileSystemDeploymentService implements DeploymentScanner, NotificationHandler {
+class FileSystemDeploymentService implements DeploymentScanner {
 
     static final Pattern ARCHIVE_PATTERN = Pattern.compile("^.*\\.(?:(?:[SsWwJjEeRr][Aa][Rr])|(?:[Ww][Aa][Bb])|(?:[Ee][Ss][Aa]))$");
 
@@ -158,70 +152,6 @@ class FileSystemDeploymentService implements DeploymentScanner, NotificationHand
     private final String relativePath;
     private final PropertyChangeListener propertyChangeListener;
     private Future<?> undeployScanTask;
-
-    @Override
-    public void handleNotification(Notification notification) {
-        if (acquireScanLock()) {
-            switch (notification.getType()) {
-                case DEPLOYMENT_DEPLOYED_NOTIFICATION: {
-                    String runtimeName = notification.getData().get(DEPLOYMENT).asString();
-                    if (!deployed.containsKey(runtimeName)) {
-                        updateDeploymentStatusAfterNotification(deploymentDir.toPath(), runtimeName);
-                    }
-                    break;
-                }
-                case DEPLOYMENT_UNDEPLOYED_NOTIFICATION: {
-                    String runtimeName = notification.getData().get(DEPLOYMENT).asString();
-                    if (deployed.containsKey(runtimeName)) {
-                        clearMarkers(deployed.get(runtimeName).parentFolder.toPath(), runtimeName);
-                        deployed.remove(runtimeName);
-                    }
-                    break;
-                }
-                default:
-                //ignore
-            }
-        }
-    }
-
-    private void updateDeploymentStatusAfterNotification(Path dir, String runtimeName) {
-        Path undeployedMarker = dir.resolve(runtimeName + UNDEPLOYED);
-        if (Files.exists(undeployedMarker)) {
-            try {
-                Files.delete(undeployedMarker);
-            } catch (IOException ioex) {
-                ROOT_LOGGER.cannotRemoveDeploymentMarker(undeployedMarker.toFile());
-            }
-            Path deployedMarker = dir.resolve(runtimeName + DEPLOYED);
-            try {
-                deployedMarker = Files.createFile(deployedMarker);
-                final Path deploymentFile = dir.resolve(runtimeName);
-                boolean isArchive = Files.exists(deploymentFile) && Files.isRegularFile(deploymentFile);
-                if (Files.exists(deploymentFile)) {
-                    Files.setLastModifiedTime(deployedMarker, Files.getLastModifiedTime(deploymentFile));
-                }
-                deployed.put(runtimeName, new DeploymentMarker(Files.getLastModifiedTime(deployedMarker).toMillis(), isArchive, dir.toFile()));
-            } catch (IOException ioex) {
-                ROOT_LOGGER.errorWritingDeploymentMarker(ioex, deployedMarker.toString());
-            }
-        }
-    }
-
-    private void clearMarkers(Path dir, String runtimeName) {
-        String fileName = runtimeName + DO_DEPLOY;
-        try {
-            Files.deleteIfExists(dir.resolve(fileName));
-            fileName = runtimeName + FAILED_DEPLOY;
-            Files.deleteIfExists(dir.resolve(fileName));
-            fileName = runtimeName + SKIP_DEPLOY;
-            Files.deleteIfExists(dir.resolve(fileName));
-            fileName = runtimeName + DEPLOYED;
-            Files.deleteIfExists(dir.resolve(fileName));
-        } catch (IOException ioex) {
-            ROOT_LOGGER.cannotRemoveDeploymentMarker(fileName);
-        }
-
-    }
 
     private class DeploymentScanRunnable implements Runnable {
 
@@ -492,7 +422,7 @@ class FileSystemDeploymentService implements DeploymentScanner, NotificationHand
 
         if (acquireScanLock()) {
             try {
-                ROOT_LOGGER.tracef("Performing a post-boot forced undeploy scan for scan directory %s", deploymentDir.getAbsolutePath());
+                ROOT_LOGGER.tracef("Performing a post-boot forced undeploy scan for scan directory %oManu", deploymentDir.getAbsolutePath());
                 ScanContext scanContext = new ScanContext(deploymentOperations);
 
                 // Add remove actions to the plan for anything we count as
@@ -1402,7 +1332,6 @@ class FileSystemDeploymentService implements DeploymentScanner, NotificationHand
             addOp.get(PERSISTENT).set(false);
             addOp.get(OWNER).set(resourceAddress);
             final ModelNode deployOp = Util.getEmptyOperation(DeploymentDeployHandler.OPERATION_NAME, address);
-            deployOp.get(OWNER).set(resourceAddress);
             return getCompositeUpdate(addOp, deployOp);
         }
 
@@ -1454,9 +1383,7 @@ class FileSystemDeploymentService implements DeploymentScanner, NotificationHand
         @Override
         protected ModelNode getUpdate() {
             final ModelNode address = new ModelNode().add(DEPLOYMENT, deploymentName);
-            final ModelNode redployOp = Util.getEmptyOperation(DeploymentRedeployHandler.OPERATION_NAME, address);
-            redployOp.get(OWNER).set(resourceAddress);
-            return redployOp;
+            return Util.getEmptyOperation(DeploymentRedeployHandler.OPERATION_NAME, address);
         }
 
         @Override
@@ -1495,7 +1422,6 @@ class FileSystemDeploymentService implements DeploymentScanner, NotificationHand
         protected ModelNode getUpdate() {
             final ModelNode address = new ModelNode().add(DEPLOYMENT, deploymentName);
             final ModelNode undeployOp = Util.getEmptyOperation(DeploymentUndeployHandler.OPERATION_NAME, address);
-            undeployOp.get(OWNER).set(resourceAddress);
             final ModelNode removeOp = Util.getEmptyOperation(DeploymentRemoveHandler.OPERATION_NAME, address);
             return getCompositeUpdate(undeployOp, removeOp);
         }
