+++ server/src/main/java/org/jboss/as/server/deployment/DeploymentAddHandler.java
@@ -139,7 +139,7 @@ public class DeploymentAddHandler implements OperationStepHandler {
         }
 
         if (ENABLED.resolveModelAttribute(context, newModel).asBoolean() && context.isNormalServer()) {
-            DeploymentHandlerUtil.deploy(context, operation, runtimeName, name, vaultReader, contentItem);
+            DeploymentHandlerUtil.deploy(context, runtimeName, name, vaultReader, contentItem);
             DeploymentUtils.enableAttribute(newModel);
         }
