+++ server/src/main/java/org/jboss/as/server/deployment/DeploymentUndeployHandler.java
@@ -44,7 +44,6 @@ public class DeploymentUndeployHandler implements OperationStepHandler {
         this.vaultReader = vaultReader;
     }
 
-    @Override
     public void execute(OperationContext context, ModelNode operation) throws OperationFailedException {
         ModelNode model = context.readResourceForUpdate(PathAddress.EMPTY_ADDRESS).getModel();
         final String deploymentUnitName = RUNTIME_NAME.resolveModelAttribute(context, model).asString();
@@ -54,7 +53,7 @@ public class DeploymentUndeployHandler implements OperationStepHandler {
         PathAddress address = PathAddress.pathAddress(opAddr);
         final String managementName = address.getLastElement().getValue();
 
-        DeploymentHandlerUtil.undeploy(context, operation, managementName, deploymentUnitName, vaultReader);
+        DeploymentHandlerUtil.undeploy(context, managementName, deploymentUnitName, vaultReader);
         DeploymentUtils.disableAttribute(model);
     }
 }
