+++ server/src/main/resources/org/jboss/as/server/controller/descriptions/LocalDescriptions.properties
@@ -236,5 +236,3 @@ deployment.enabled-time=Last time the application was enabled
 deployment.enabled-timestamp=Last timestamp the application was enabled. Format is yyyy-MM-dd HH:mm:ss,SSS zzz.
 deployment.disabled-time=Last time the application was disabled
 deployment.disabled-timestamp=Last timestamp the application was disabled. Format is yyyy-MM-dd HH:mm:ss,SSS zzz.
-deployment.deployment-deployed=Notification sent when a deployment is deployed.
-deployment.deployment-undeployed=Notification sent when a deployment is undeployed.
