+++ testsuite/manualmode/src/test/java/org/jboss/as/test/manualmode/deployment/DeploymentScannerUnitTestCase.java
@@ -19,28 +19,40 @@
 package org.jboss.as.test.manualmode.deployment;
 
 import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.DEPLOYMENT;
+import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.EXTENSION;
+import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.INCLUDE_RUNTIME;
+import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.NAME;
 import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.OUTCOME;
+import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.READ_ATTRIBUTE_OPERATION;
+import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.READ_RESOURCE_OPERATION;
+import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.RECURSIVE;
+import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.RESULT;
+import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.SUBSYSTEM;
+import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.SUCCESS;
+import static org.jboss.as.test.integration.management.util.ModelUtil.createOpNode;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
 import java.io.File;
 import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.UUID;
 
 import javax.inject.Inject;
 
 import org.apache.commons.io.FileUtils;
 import org.jboss.as.controller.PathAddress;
+import org.jboss.as.controller.PathElement;
 import org.jboss.as.controller.client.ModelControllerClient;
 import org.jboss.as.controller.descriptions.ModelDescriptionConstants;
 import org.jboss.as.controller.operations.common.Util;
 import org.jboss.as.protocol.StreamUtils;
-import org.jboss.as.server.deployment.DeploymentUndeployHandler;
 import org.jboss.as.test.shared.TestSuiteEnvironment;
 import org.jboss.as.test.shared.TimeoutUtil;
 import org.jboss.dmr.ModelNode;
+import org.jboss.shrinkwrap.api.ShrinkWrap;
+import org.jboss.shrinkwrap.api.asset.StringAsset;
+import org.jboss.shrinkwrap.api.exporter.ZipExporter;
+import org.jboss.shrinkwrap.api.spec.JavaArchive;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
@@ -55,7 +67,7 @@ import org.wildfly.core.testrunner.WildflyTestRunner;
  */
 @RunWith(WildflyTestRunner.class)
 @ServerControl(manual = true)
-public class DeploymentScannerUnitTestCase extends AbstractDeploymentUnitTestCase {
+public class DeploymentScannerUnitTestCase {
 
     private static final PathAddress DEPLOYMENT_ONE = PathAddress.pathAddress(DEPLOYMENT, "deployment-one.jar");
     private static final PathAddress DEPLOYMENT_TWO = PathAddress.pathAddress(DEPLOYMENT, "deployment-two.jar");
@@ -97,7 +109,7 @@ public class DeploymentScannerUnitTestCase extends AbstractDeploymentUnitTestCas
                 createDeployment(deploymentTwo, "non.existing.dependency");
 
                 // Add a new de
-                addDeploymentScanner(0);
+                addDeploymentScanner();
                 try {
                     // Wait until deployed ...
                     long timeout = System.currentTimeMillis() + TimeoutUtil.adjust(30000);
@@ -109,8 +121,8 @@ public class DeploymentScannerUnitTestCase extends AbstractDeploymentUnitTestCas
                     Assert.assertTrue(exists(DEPLOYMENT_TWO));
                     Assert.assertEquals("FAILED", deploymentState(DEPLOYMENT_TWO));
 
-                    final Path oneDeployed = deployDir.toPath().resolve("deployment-one.jar.deployed");
-                    final Path twoFailed =  deployDir.toPath().resolve("deployment-two.jar.failed");
+                    final File oneDeployed = new File(deployDir, "deployment-one.jar.deployed");
+                    final File twoFailed = new File(deployDir, "deployment-two.jar.failed");
 
                     // Restart ...
                     container.stop();
@@ -122,8 +134,8 @@ public class DeploymentScannerUnitTestCase extends AbstractDeploymentUnitTestCas
                         Thread.sleep(200);
                     }
 
-                    Assert.assertTrue(Files.exists(oneDeployed));
-                    Assert.assertTrue(Files.exists(twoFailed));
+                    Assert.assertTrue(oneDeployed.exists());
+                    Assert.assertTrue(twoFailed.exists());
 
                     Assert.assertTrue(exists(DEPLOYMENT_ONE));
                     Assert.assertEquals("OK", deploymentState(DEPLOYMENT_ONE));
@@ -133,22 +145,7 @@ public class DeploymentScannerUnitTestCase extends AbstractDeploymentUnitTestCas
                         Thread.sleep(200);
                     }
                     Assert.assertFalse(exists(DEPLOYMENT_TWO));
-                    ModelNode disableScanner = Util.getWriteAttributeOperation(PathAddress.parseCLIStyleAddress("/subsystem=deployment-scanner/scanner=testScanner"), "scan-interval", 300000);
-                    ModelNode result = executeOperation(disableScanner);
-                    assertEquals("Unexpected outcome of disabling the test deployment scanner: " + disableScanner, ModelDescriptionConstants.SUCCESS, result.get(OUTCOME).asString());
 
-                    final ModelNode undeployOp = Util.getEmptyOperation(DeploymentUndeployHandler.OPERATION_NAME, DEPLOYMENT_ONE.toModelNode());
-                    result = executeOperation(undeployOp);
-                    assertEquals("Unexpected outcome of undeploying deployment one: " + undeployOp, ModelDescriptionConstants.SUCCESS, result.get(OUTCOME).asString());
-                    Assert.assertTrue(exists(DEPLOYMENT_ONE));
-                    Assert.assertEquals("STOPPED", deploymentState(DEPLOYMENT_ONE));
-
-                    timeout = System.currentTimeMillis() + TimeoutUtil.adjust(10000);
-
-                    while ( Files.exists(oneDeployed) && System.currentTimeMillis() < timeout) {
-                        Thread.sleep(10);
-                    }
-                    Assert.assertFalse(Files.exists(oneDeployed));
                 } finally {
                     removeDeploymentScanner();
                 }
@@ -161,14 +158,105 @@ public class DeploymentScannerUnitTestCase extends AbstractDeploymentUnitTestCas
         }
     }
 
+    private void addDeploymentScanner() throws Exception {
+        ModelNode addOp = Util.createAddOperation(PathAddress.pathAddress(PathElement.pathElement(EXTENSION, "org.jboss.as.deployment-scanner")));
+        ModelNode result = executeOperation(addOp);
+        assertEquals("Unexpected outcome of adding the test deployment scanner extension: " + addOp, ModelDescriptionConstants.SUCCESS, result.get("outcome").asString());
+        addOp = Util.createAddOperation(PathAddress.pathAddress(PathElement.pathElement(SUBSYSTEM, "deployment-scanner")));
+        result = executeOperation(addOp);
+        assertEquals("Unexpected outcome of adding the test deployment scanner subsystem: " + addOp, ModelDescriptionConstants.SUCCESS, result.get("outcome").asString());
+        // add deployment scanner
+        final ModelNode op = getAddDeploymentScannerOp();
+        result = executeOperation(op);
+        assertEquals("Unexpected outcome of adding the test deployment scanner: " + op, ModelDescriptionConstants.SUCCESS, result.get("outcome").asString());
+    }
 
-    @Override
-    protected ModelNode executeOperation(ModelNode op) throws IOException {
+    private void removeDeploymentScanner() throws Exception {
+        boolean ok = false;
+        try {
+            // remove deployment scanner
+            final ModelNode op = getRemoveDeploymentScannerOp();
+            ModelNode result = executeOperation(op);
+            assertEquals("Unexpected outcome of removing the test deployment scanner: " + op, ModelDescriptionConstants.SUCCESS, result.get("outcome").asString());
+            ok = true;
+        } finally {
+            try {
+                boolean wasOK = ok;
+                ok = false;
+                ModelNode removeOp = Util.createRemoveOperation(PathAddress.pathAddress(PathElement.pathElement(SUBSYSTEM, "deployment-scanner")));
+                ModelNode result = executeOperation(removeOp);
+                if (wasOK) {
+                    assertEquals("Unexpected outcome of removing the test deployment scanner subsystem: " + removeOp, ModelDescriptionConstants.SUCCESS, result.get("outcome").asString());
+                } // else don't override the previous assertion error in this finally block
+                ok = wasOK;
+            } finally {
+                ModelNode removeOp = Util.createRemoveOperation(PathAddress.pathAddress(PathElement.pathElement(EXTENSION, "org.jboss.as.deployment-scanner")));
+                ModelNode result = executeOperation(removeOp);
+                if (ok) {
+                    assertEquals("Unexpected outcome of removing the test deployment scanner extension: " + removeOp, ModelDescriptionConstants.SUCCESS, result.get("outcome").asString());
+                }  // else don't override the previous assertion error in this finally block
+            }
+        }
+    }
+
+    private ModelNode executeOperation(ModelNode op) throws IOException {
         return client.execute(op);
     }
 
-    @Override
-    protected File getDeployDir() {
-        return deployDir;
+    private ModelNode getAddDeploymentScannerOp() {
+        final ModelNode op = Util.createAddOperation(getTestDeploymentScannerResourcePath());
+        op.get("scan-interval").set(0);
+        op.get("path").set(deployDir.getAbsolutePath());
+        return op;
     }
+
+    private ModelNode getRemoveDeploymentScannerOp() {
+        return createOpNode("subsystem=deployment-scanner/scanner=testScanner", "remove");
+    }
+
+    private PathAddress getTestDeploymentScannerResourcePath() {
+        return PathAddress.pathAddress(PathElement.pathElement("subsystem", "deployment-scanner"), PathElement.pathElement("scanner", "testScanner"));
+    }
+
+    protected void createDeployment(final File file, final String dependency) throws IOException {
+        final JavaArchive archive = ShrinkWrap.create(JavaArchive.class);
+        final String dependencies = "Dependencies: " + dependency;
+        archive.add(new StringAsset(dependencies), "META-INF/MANIFEST.MF");
+        archive.as(ZipExporter.class).exportTo(file);
+    }
+
+    protected boolean exists(PathAddress address) throws IOException {
+        final ModelNode operation = Util.createEmptyOperation(READ_RESOURCE_OPERATION, address);
+        operation.get(INCLUDE_RUNTIME).set(true);
+        operation.get(RECURSIVE).set(true);
+
+        final ModelNode result = executeOperation(operation);
+        if (SUCCESS.equals(result.get(OUTCOME).asString())) {
+            return true;
+        }
+        return false;
+    }
+
+    protected String deploymentState(final PathAddress address) throws IOException {
+        final ModelNode operation = Util.createEmptyOperation(READ_ATTRIBUTE_OPERATION, address);
+        operation.get(NAME).set("status");
+
+        final ModelNode result = executeOperation(operation);
+        if (SUCCESS.equals(result.get(OUTCOME).asString())) {
+            return result.get(RESULT).asString();
+        }
+        return "failed";
+    }
+
+    protected boolean isRunning() throws IOException {
+        final ModelNode operation = Util.createEmptyOperation(READ_ATTRIBUTE_OPERATION, PathAddress.EMPTY_ADDRESS);
+        operation.get(NAME).set("server-state");
+
+        final ModelNode result = executeOperation(operation);
+        if (SUCCESS.equals(result.get(OUTCOME).asString())) {
+            return "running".equals(result.get(RESULT).asString());
+        }
+        return false;
+    }
+
 }
